requirements.txt
mssql-django 
python-dotenv
Pillow

